Remove the queryparam ?beta=true in perspectives app (#397)
* Replaced perspective folders with new perspectives folder and updated loadView

* updated module paths
