+++ tests/view/perspectives/loadPerspectivePicker.js
@@ -17,7 +17,7 @@ chai.use(chaiAsPromised);
 const expect = chai.expect;
 import { statuses } from '../../../api/v1/constants';
 import sinon from 'sinon';
-import { getValuesObject } from '../../../view/newPerspective/utils.js';
+import { getValuesObject } from '../../../view/perspective/utils.js';
 
 describe('Perspective app ', () => {
   const tags = ['one', 'two', 'three', 'four', 'five'];
