+++ tests/view/perspectives/perspectiveController.js
@@ -13,7 +13,7 @@
 import { expect } from 'chai';
 import React from 'react';
 import sinon from 'sinon';
-import PerspectiveController from '../../../view/newPerspective/PerspectiveController.js';
+import PerspectiveController from '../../../view/perspective/PerspectiveController.js';
 import { mount } from 'enzyme';
 
 describe('Perspective controller ', () => {
