+++ view/loadView.js
@@ -133,13 +133,7 @@ module.exports = function loadView(app, passport) {
         // updates
         if ((key === '/perspectives' && Object.keys(req.query).length) ||
         key === '/perspectives/:key') {
-
-          // if ?beta=true in URL, use the new perspective loader
-          if (req.query.beta) {
-            res.render('newPerspective/perspective', templateVars);
-          } else {
-            res.render(viewmap[key], templateVars);
-          }
+          res.render(viewmap[key], templateVars);
         } else {
           res.render(viewmap[key], trackObj);
         }
