--- view/newPerspective/perspective.pug
@@ -1,40 +1,0 @@
-//
- Copyright (c) 2016, salesforce.com, inc.
- All rights reserved.
- Licensed under the BSD 3-Clause license.
- For full license text, see LICENSE.txt file in the repo root or
- https://opensource.org/licenses/BSD-3-Clause
-
-
-doctype html
-html(lang = 'en')
-  head
-    link(rel='shortcut icon', href='favicon.ico')
-    link(rel='stylesheet', type='text/css', href='../static/css/salesforce-lightning-design-system.min.css')
-    link(rel='stylesheet' type='text/css' href='../static/css/perspective.css')
-    title Refocus
-
-  body
-    #refocus_perspective_dropdown_container
-    #create_perspective_container
-    label#errorInfo
-    div#lens
-    div#lens_loading_spinner
-      .slds-spinner_container
-        .slds-spinner--brand.slds-spinner.slds-spinner--medium(role='alert')
-          span.slds-assistive-text Loading
-          .slds-spinner__dot-a
-          .slds-spinner__dot-b
-
-  script(src='/static/socket.io.js')
-  script.
-    var realtimeEventThrottleMilliseconds = #{eventThrottle};
-  script.
-    var transProtocol = '#{transportProtocol}';
-  script
-    if queryParams
-      | var queryParams = !{queryParams}
-  script(src='/static/newPerspective/app.js')
-  script.
-    var trackingId = '#{trackingId}';
-  script(src='../static/analytics/app.js')
