+++ view/perspective/utils.js
@@ -48,7 +48,7 @@ function getFilterQuery(p) {
         p.subjectTagFilter.join().replace(/,/g, ',' + sign);
   }
 
-  if (p.statusFilter.length) {
+  if (p.statusFilter && p.statusFilter.length) {
     if (q) {
       q += '&';
     }
@@ -77,11 +77,8 @@ function getArray(field, arrayOfObjects) {
   }
 
   for (let i = 0; i < arrayOfObjects.length; i++) {
-    if (arrayOfObjects[i].isPublished) {
-      arr.push(arrayOfObjects[i][field]);
-    }
+    arr.push(arrayOfObjects[i][field]);
   }
-
   return arr;
 }
 
@@ -112,9 +109,7 @@ function convertCamelCase(string) {
  * @returns {Array} The array of strings or primitives
  */
 function filteredArray(arr, removeThis) {
-  return arr.filter((elem) => {
-    return elem !== removeThis;
-  });
+  return arr.filter((elem) => elem && elem !== removeThis);
 }
 
 /**
@@ -176,9 +171,9 @@ function getConfig(values, key, value) {
   };
 
   if (key === 'subjects') {
+    config.placeholderText = 'Select a Subject...';
     let options = getArray('absolutePath', values[key]);
     config.options = filteredArray(options, value);
-    config.placeholderText = 'Select a Subject...';
     config.isArray = false;
   } else if (key === 'lenses') {
     config.placeholderText = 'Select a Lens...';
@@ -198,7 +193,7 @@ function getConfig(values, key, value) {
       config.allOptionsLabel = 'All ' +
         convertedText.replace(' Filter', '') + 's';
       let options = getArray('name', values[key]);
-      config.options = getOptions(options, value);
+      config.options = filteredArray(values[key], value);
     }
 
     delete config.placeholderText;
@@ -207,11 +202,222 @@ function getConfig(values, key, value) {
   return config;
 }
 
-module.exports = {
+/**
+ * Return array of unique tags
+ * @param {Array} Objects with tags: [tag1, tag2, ...]
+ * @returns {Array} contains unique tags
+ */
+function getTagsFromArrays(arr) {
+  let tags = new Set();
+  for (let i = arr.length - 1; i >= 0; i--) {
+    if (arr[i] && arr[i].tags && arr[i].tags.length) {
+      tags = new Set([...tags, ...new Set(arr[i].tags)]);
+    }
+  }
+
+  return [...tags].sort();
+}
+
+/**
+ * Return only the published objects.
+ * @param {Array} the array to filter from.
+ * Contains objects with key isPublished
+ * @returns {Array} with only objects containing isPublished: true
+*/
+function getPublishedFromArr(arr) {
+  return arr.filter((elem) => elem.isPublished);
+}
+
+/**
+ * Accumulates information to load the perspective dropdown,
+ * and the edit/create perspective modal
+ *
+ * @param {Object} accumulatorObject
+ * @returns {Object} The accumulated values
+ */
+function getValuesObject(accumulatorObject) {
+  const {
+    getPromiseWithUrl,
+    getPerspectiveUrl,
+    handleHierarchyEvent,
+    handleLensDomEvent,
+    customHandleError,
+  } = accumulatorObject;
+  const constants = require('../../api/v1/constants');
+  const httpStatus = constants.httpStatus;
+  const statuses = constants.statuses;
+
+  const valuesObj = {
+    name: '',
+    perspectives: [], // will be array of objects
+    perspective: null, // if found becomes object
+    persNames: [], // will be array of strings
+    rootSubject: {},
+    lens: {}, // includes library
+    subjects: [], // { name: absolutePath, id }
+    aspectTagFilter: [], // { name, id }
+    aspectFilter: [], // strings
+    subjectTagFilter: [], // strings
+    lenses: [], // { name, id }
+    statusFilter: Object.keys(statuses).sort(),
+  };
+  let hierarchyLoadEvent; // will be custom event
+  let gotLens = false;
+
+  // if perspective is named, get the named perspective
+  // otherwise get the default perspective
+  const { url, named } = getPerspectiveUrl();
+
+  function getPageLoadingPromises(perspective) {
+    /*
+     * getLens and getHierarchy are dispatched simultaneously.
+     * Both lensLoadEvent and hierarchyLoadEvent will be called once:
+     *
+     * If getLens resolves first,
+     * 1. In handleLensDomEvent:
+     * hierarchyLoadEvent is NOT be dispatched since it is not defined,
+     * lensLoadEvent is dispatched.
+     * 2. gotLens is set to true.
+     * 3. When getHierarchy resolves:
+     * in handleHierarchyEvent, because gotLens is true, hierarchyLoadEvent
+     * is dispatched.
+     * Else if getHierarchy is resolved first:
+     * 1. in handleHierarchyEvent:
+     * Because gotLens is false,
+     * hierarchyLoadEvent is NOT dispatched. Instead it is returned and assigned
+     * to the variable hierarchyLoadEvent.
+     * 2. When getLens resolves, in handleLensDomEvent:
+     * lensLoadEvent is dispatched. Since hierarchyLoadEvent is truthy,
+     * it is also dispatched.
+     */
+    const getLens = getPromiseWithUrl('/v1/lenses/' + perspective.lensId)
+    .then((res) => {
+
+      // hierarchyLoadEvent can be undefined or a custom event
+      // if hierarchyLoadEvent is custom event, it will be dispatched
+      handleLensDomEvent(res.body.library, hierarchyLoadEvent);
+
+      // set the lens received flag to true, to dispatch lens load
+       // when hierarchy is resolved in getHierarchy
+      gotLens = true;
+
+      valuesObj.lens = res.body;
+    });
+
+    const filterString  = getFilterQuery(perspective);
+    const getHierarchy = getPromiseWithUrl('/v1/subjects/' +
+    perspective.rootSubject + '/hierarchy' + filterString)
+    .then((res) => {
+
+      // if gotLens is false, hierarchyLoadEvent will be assigned
+      // and NOT dispatched. Otherwise dispatch the hierarchy event
+      hierarchyLoadEvent = handleHierarchyEvent(res.body, gotLens);
+
+      valuesObj.rootSubject = res.body;
+    });
+
+    return [getLens, getHierarchy];
+  }
+
+  // Need to get all perspectives, for editing perspectives.
+  // for GET perspective, do NOT throw error if 404, since
+  // throwing an error skips loading the perspective picker.
+  // If successful, load the hierarchy and lens
+  const arr = [
+    getPromiseWithUrl('/v1/perspectives'),
+    getPromiseWithUrl(url)
+    .then((res) => {
+      Promise.all(getPageLoadingPromises(res.body));
+      return res;
+    })
+    .catch(console.log)
+  ];
+
+  return Promise.all(arr)
+  .then((responses) => {
+    valuesObj.perspectives = responses[0].body;
+
+    // use ternary as if pespective does not exist, responses[1] is undefined
+    const returnedPerspective = responses[1] ? responses[1].body : null;
+
+    // assign perspective-related values to the accumulator object
+    valuesObj.persNames = valuesObj.perspectives
+      .map((perspective) => perspective.name).sort();
+
+    /*
+     * One out of four situations can happen:
+     * GET named perspective: exists. Assign perspective
+     * GET default perspective: exists. Assign perspective
+     * GET named perspective: does NOT exist: handleError
+     * GET default perspective: does NOT exist: perspective is the first
+     *  perspective in the perspectives array. If no perspectives exist,
+     *  valuesObj.perspective = null
+     */
+     if (returnedPerspective) {
+        valuesObj.perspective = returnedPerspective;
+        valuesObj.name = valuesObj.perspective.name;
+     } else if (named) {
+
+        // named perspective does not exist
+        const name = url.split('/').pop();
+        customHandleError('Sorry, but the perspective you were trying ' +
+          'to load, ' + name + ', does not exist. Please select a ' +
+          'perspective from the dropdown.');
+      } else {
+
+        // default perspective does NOT exist.
+        // GET the first perspective by alphabetical order.
+        // Check to see there are perspectives
+        if (valuesObj.perspectives.length) {
+
+          // redirect to the first perspective. The rest of the code
+          // won't be executed.
+          window.location.href = '/perspectives/' + valuesObj.perspectives[0].name;
+        } else {
+          valuesObj.perspective = null;
+
+          // no perspectives exist
+          customHandleError('no perspectives exist.');
+      }
+    }
+
+    // valuesObj.perspective have been assigned.
+    const promisesArr = [
+      getPromiseWithUrl('/v1/lenses?fields=isPublished,name'),
+      getPromiseWithUrl('/v1/subjects?fields=isPublished,absolutePath,tags'),
+      getPromiseWithUrl('/v1/aspects?fields=isPublished,name,tags'),
+    ];
+
+    return Promise.all(promisesArr);
+  })
+  .then((responses) => {
+    const lenses = responses[0].body;
+    const subjects = responses[1].body;
+    const aspects = responses[2].body;
+
+    // assign non-perspective values to the accumulator object.
+    // TODO: subjects are objects. Change to use strings
+    valuesObj.subjects = getPublishedFromArr(subjects);
+    valuesObj.subjectTagFilter = getTagsFromArrays(valuesObj.subjects);
+    valuesObj.lenses = getPublishedFromArr(lenses);
+
+    // aspectFilter is an array of strings
+    const publishedAspects = getPublishedFromArr(aspects);
+    valuesObj.aspectFilter = publishedAspects.map((aspect) => aspect.name);
+    valuesObj.aspectTagFilter = getTagsFromArrays(publishedAspects);
+
+    return valuesObj;
+  });
+} // getValuesObject
+
+module.exports =  {
+  getValuesObject,
+  getTagsFromArrays,
   getFilterQuery,
   getOptions, // for testing
   filteredArray,
   getConfig,
   getArray,
   getTagsFromResources,
+  getPublishedFromArr,
 };
